[sailfishos][gecko] Allow LoginManagerPrompter to find its window. JB#55760, OMP#JOLLA-418
This patch blocks loading of gecko's LoginManagerAuthPrompter.jsm so
that the embedlite-components version can be used instead.

It also patches the nsILoginManagerPrompter interface to allow a
reference to the window to be passed through, to allow the embedlite
component to understand its context.

Finally it patches ParentChannelListener to pass the correct window
object through to the nsILoginManagerAuthPrompter component.

Note: this is the application of patch 0082.