Drop debuginfo for Rust components
The style component fails to build on aarch64 when debuginfo is enabled.
The build fails with memory errors and the following:

fatal runtime error: Rust cannot catch foreign exceptions

Disabling debuginfo while building this file allows it to go through.

This change drops debuginfo for all Rust components, so it may be better
to use a more targeted approach instead.