Work around upstream membarrier changes
Upstream changes for Bugzilla bug 1661016 "aarch64: Invalidate icache
when compiling on a background thread" introduce the use of membarrier.
See:

https://phabricator.services.mozilla.com/D88395

However, Linux kernel headers 4.16 or more recent are required for the
membarrier.h file and related syscalls to be present. Since our headers
are being held back 3.18 we don't have access to these.

https://github.com/sailfishos/kernel-headers

This change drops the use of the header file and defines
__NR_membarrier. It's not clear what the runtime consequences of this
will be, but if this causes problems we should consider reverting
changeset D88395 instead.

The problems this fixes seem to only cause trouble in aarch64 simulation
environments.

For reference:

https://man7.org/linux/man-pages/man2/membarrier.2.html