Supress URLQueryStrippingListService.jsm error
The URLQueryStrippingListService service is initialised by the
"profile-after-change" event. The event also goes through to the
observer, which is coded to generate an error when the event is received
("unexpected event profile-after-change"). This change supressed the
error as long as the event is received before the service has been
initialised (i.e. it's the event that's caused the service to be
initialised).

It's not clear why this isn't an issue upstream.

See the following for context:

https://bugzilla.mozilla.org/show_bug.cgi?id=1706608

https://phabricator.services.mozilla.com/D117376

https://bugzilla.mozilla.org/show_bug.cgi?id=1723981