Allow file scheme when loading OpenSearch providers
When loading OpenSearch provider XML files the scheme was restricted to
https. This prevented the OpenSearch XML provider files from being
loaded from disk at initialisation or when installing a search provider.

This change extends the support to include both https and file schemes
so that the providers can be loaded correctly from disc.