Drop AudioPlayback messages if no embedder element is defined
The AudioPlayback Window Actor passes "audio-playback" messages on to
the embedder element by calling it directly. If the embedder element is
null this will result in a "browser is null" error being output. This
doesn't do any real harm, but is unnecessary and could cause confusion.
This change adds a null check so that an attempt to call methods off the
browser only happen if it's defined.

On Sailfish OS the embedder element is always null, so there's no need
(or way) to pass these calls on anyway.