Set security.external_protocol_requires_permission to default to false
Upstream introduces the new user preference:

security.external_protocol_requires_permission

This defaults to true, but we need it to default to false so that
external URL actions continue to work as intended. Here we change the
default value of the preference from true to false so as to allow this.