Set embedlite.azpc.json.doubletap static pref to default to true
When the preferences were moved to the StaticPrefList.yaml file the
embedlite.azpc.json.doubletap was set to default to false, but it should
have been set to true. If set to false, double-tap always zooms, even on
items that support double-taps.

This should be combined with commit e035d6ff3a785 "Add embedlite static
prefs".