Create GLLibraryEGL as a singleton in GLContextProviderEGL
Ensure that if GLLibraryEGL is creaetd through the GLContextProviderEGL
interface then only one instance is ever created. This ensures that the
internal display management works as expected and avoids a crash that
can occur when a display is freed and the fTerminate method is called.

This should be combined with commit 1765dd0460a9032c "Restore
GLScreenBuffer and TextureImageEGL".